moved testSubmoduleUpdate to GitAPITestUpdateCliGit
